Prepare for push notifications
Constantly poll for prices instead of on each request. Now requests will
receive the latest cached price information. Prices are fed into a
diffing function which for now just logs when the difference is higher
than the previous update.

This is in preparation for pushing notifications via Twilio, which will
warrant smarter rules around when to send and how to group
notifications.