Add Huobi
Adds prices for Huobi exchange. Note that Huobi does not provide an
endpoint (at least that is published) that returns all prices. This
means we have to make a request per coin instead of a single request.
They also do not provide the last deal price in the market detail API,
so the highest bid will be displayed as the Huobi price.