Make mobile alerts generic
Updates the diff checker to accept a list of exchanges instead of
hardcoding Bittrex and Binance. The list can then accept an arbitrary
number of exchanges as long as the price data is formatted.

Kucoin is hardcoded to only notify for certain thresholds since it has
low liquidity for most coins.

This change in effect adds Huobi and Kucoin to mobile diff alerts.