Don't hardcode shared coins on the server
Programmatically determines what coins are shared between more than one
exchange using the list of coins on each exchange (these lists are still
hardcoded). Note that the front-end still has a hardcoded list of shared
coins to maintain a certain ordering (in the future this can be done
programmatically and the page jumping will be fine since it should only
happen when a new coin is added that is shared between multiple
exchanges).