Make notifications stricter for HitBTC/Kucoin
HitBTC has low liquidity for many coins as well as large price steps,
leading many diffs to be basically unactionable. This ups the required
diff in order for a notification to be sent for HitBTC (sets an even
higher requirement if the diff is between HitBTC and Kucoin, another
low liquidity exchange).

TO DO
Use available orders/volume to calculate and notify based on potential
net profit, not just the straight diff (since this diff may only apply
for a very small volume).