Show last received Coinbase price instead of NaN
Coinbase rate limits its API and we were previously showing NaN for
Coinbase prices whenever ticker data was not receieved. This changes
the behavior to keep displaying the last received price, but
highlighting that price in red to indicate that the information is
outdated. The next successful response will remove the highlighting.